Wired up Recycler View
I wired up databinding and domain model to database model class conversion for the Asteroids.

The rest of the problem has to do with getting relevant data back into the room database as I had to destroy and increment the version due to schema changes which resulted in the destruction of all data.

This is a good thing because it allows an opportunity to ensure our refresh methods for when a device is charging works as expected.

I will be live loading this on a product tomrorrow for more testing/debugging.