Filtering Logic
I setup three public properties getAsteroids, getWeeklyAsteroids, and getTodayAsteroids which bind to the action bar in onOptionSelected to determine when those filters should be applied.

The issue is I have getting null responses from the database on all three of them.

I first suspected my type converter being the problem but now I am curious if there is an issue with the casting that is occurring in the view model.

I will be working on this next.